Improve user deletion control checking available deposit

+++ src/main/java/com/machines/vending/domain/commands/user/DeleteUserCommandImpl.java
@@ -1,5 +1,8 @@
 package com.machines.vending.domain.commands.user;
 
+import com.machines.vending.domain.commands.deposit.ReadDepositCommand;
+import com.machines.vending.domain.exceptions.PositiveDepositAvailableException;
+import com.machines.vending.domain.models.Deposit;
 import com.machines.vending.domain.models.User;
 import com.machines.vending.infrastructure.persistence.repositories.UserRepository;
 import lombok.AllArgsConstructor;
@@ -9,10 +12,14 @@ import org.springframework.stereotype.Service;
 @AllArgsConstructor
 public class DeleteUserCommandImpl implements DeleteUserCommand {
     private final UserRepository userRepository;
+    private final ReadDepositCommand readDepositCommand;
 
     @Override
-    public void execute(final User user) {
-        // TODO Check deposit
+    public void execute(final User user) throws PositiveDepositAvailableException {
+        int depositAmount = readDepositCommand.read(Deposit.builder().buyerId(user.getId()).build()).getAmount();
+        if (depositAmount > 0) {
+            throw new PositiveDepositAvailableException();
+        }
         userRepository.deleteById(user.getId());
     }
 }
